+++ samples/DotNetCoreSample/DependencyInjectionTest.cs
@@ -273,11 +273,35 @@ namespace DotNetCoreSample
         void Fly();
 
         void OpenFly<T>();
+
+        string FlyAway();
     }
 
     public interface IAnimal<T>
     {
         void Eat();
+    }
+
+    public class Animal<T>
+    {
+        private int _eatCounter;
+        private int _drinkCounter;
+
+        public virtual void Eat()
+        {
+            Console.WriteLine("Eating now");
+            _eatCounter++;
+        }
+
+        public int GetEatCount() => _eatCounter;
+
+        public virtual void Drink(T t)
+        {
+            Console.WriteLine($"ddd {t}");
+            _drinkCounter++;
+        }
+
+        public virtual int GetDrinkCount() => _drinkCounter;
     }
 
     public class LogInterceptor : AbstractInterceptor
@@ -312,6 +336,8 @@ namespace DotNetCoreSample
             Console.WriteLine($"{Name} is OpenFlying,OpenType: {typeof(T).FullName}");
         }
 
+        public string FlyAway() => "十万八千里";
+
         public void Dispose()
         {
             Console.WriteLine($"{Name}  disposed...");
@@ -331,6 +357,8 @@ namespace DotNetCoreSample
         {
             Console.WriteLine($"{Name} is OpenFlying,OpenType: {typeof(T).FullName}");
         }
+
+        public string FlyAway() => "xxxxx";
     }
 
     internal class TestServiceContainerModule : ServiceContainerModule