+++ src/WeihanLi.Common/Aspect/AspectDelegate.cs
@@ -8,9 +8,24 @@ namespace WeihanLi.Common.Aspect
 {
     public class AspectDelegate
     {
+        public static void Invoke(IInvocation context)
+        {
+            InvokeInternal(context, null, null);
+        }
+
         public static void InvokeWithInterceptors(IInvocation invocation, IReadOnlyCollection<IInterceptor> interceptors)
         {
-            var action = GetAspectDelegate(invocation, interceptors);
+            InvokeInternal(invocation, interceptors, null);
+        }
+
+        public static void InvokeWithCompleteFunc(IInvocation invocation, Func<IInvocation, Task> completeFunc)
+        {
+            InvokeInternal(invocation, null, completeFunc);
+        }
+
+        public static void InvokeInternal(IInvocation invocation, IReadOnlyCollection<IInterceptor> interceptors, Func<IInvocation, Task> completeFunc)
+        {
+            var action = GetAspectDelegate(invocation, interceptors, completeFunc);
             var task = action.Invoke(invocation);
             if (!task.IsCompleted)
             {
@@ -34,21 +49,24 @@ namespace WeihanLi.Common.Aspect
             }
         }
 
-        private static Func<IInvocation, Task> GetAspectDelegate(IInvocation invocation, IReadOnlyCollection<IInterceptor> interceptors)
+        private static Func<IInvocation, Task> GetAspectDelegate(IInvocation invocation, IReadOnlyCollection<IInterceptor> interceptors, Func<IInvocation, Task> completeFunc)
         {
             // ReSharper disable once ConvertToLocalFunction
-            Func<IInvocation, Task> completeFunc = x =>
+            // ReSharper disable once ConvertIfStatementToNullCoalescingAssignment
+            if (null == completeFunc)
             {
-                invocation.ReturnValue = x.Method?.Invoke(x.Target, x.Arguments);
-                if (invocation.ProxyMethod.ReturnType == typeof(void))
+                completeFunc = x =>
                 {
-                    return TaskHelper.CompletedTask;
-                }
+                    invocation.ReturnValue = x.Method?.Invoke(x.Target, x.Arguments);
+                    if (invocation.ProxyMethod.ReturnType == typeof(void))
+                    {
+                        return TaskHelper.CompletedTask;
+                    }
 
-                if (invocation.ReturnValue is Task task)
-                {
-                    return task;
-                }
+                    if (invocation.ReturnValue is Task task)
+                    {
+                        return task;
+                    }
 #if NETSTANDARD2_1
                 if (invocation.ReturnValue is ValueTask valTask)
                 {
@@ -56,8 +74,9 @@ namespace WeihanLi.Common.Aspect
                 }
 #endif
 
-                return TaskHelper.CompletedTask;
-            };
+                    return TaskHelper.CompletedTask;
+                };
+            }
 
             // ReSharper disable once ConvertIfStatementToNullCoalescingAssignment
             if (null == interceptors)
@@ -79,11 +98,6 @@ namespace WeihanLi.Common.Aspect
             }
 
             return builder.Build();
-        }
-
-        public static void Invoke(IInvocation context)
-        {
-            InvokeWithInterceptors(context, null);
         }
     }
 }